services/backend/app/requirements.txt
fastapi
fastapi-crudrouter
SQLAlchemy
SQLAlchemy-Utils
psycopg2-binary
python-jose
passlib
uvicorn
psycopg2cffi
python-multipart